[feat]型定義をtypes配下にまとめた #14
