+++ src/data/project.ts
@@ -1,26 +1,5 @@
 // src/data/works.ts
-export interface Project {
-  id: string;
-  author: string;
-  title: string;
-  createdAt: string; // 作成日
-  updatedAt: Date;
-}
-
-// interface Row {
-//   stitches: Stitch[];        // この行の目（Stitch）の配列
-// }
-
-export type Language = "en" | "ja";
-
-// interface Stitch {
-//   stitchType: "slip" | "chain" | "single" | "double" | "treble" | string;  // 目の種類
-//   isCompleted: boolean;      // この目が完了しているか
-//   isMarked: boolean;         // ユーザーがマークしたかどうか
-//   comment?: string;             // コメント
-//   markerColor?: string;      // メモの色（任意）
-//   rowId: number;             // 目が所属する行のID
-// }
+import { Project } from "../types/Project.types";
 
 export const dummyProjects: Project[] = [
   {
